requirements.txt
Flask
requests
PyGithub
scikit-learn
pandas
python-dotenv